refactor(MVC): refactoring files to follow MVC
