--- src/main/kotlin/lotto/Controller.kt
@@ -1,5 +1,0 @@
-package lotto
-
-fun main() {
-    LottoMachine.start()
-}
