+++ src/main/kotlin/lotto/controller/Controller.kt
@@ -1,0 +1,7 @@
+package lotto.controller
+
+import lotto.model.LottoMachine
+
+fun main() {
+    LottoMachine.start()
+}
