+++ src/test/kotlin/lotto/AmountValidatorTest.kt
@@ -1,5 +1,7 @@
 package lotto
 
+import lotto.service.AmountValidator
+import lotto.service.Constants
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertDoesNotThrow
