+++ src/test/kotlin/lotto/LottoNumberTest.kt
@@ -1,5 +1,7 @@
 package lotto
 
+import lotto.model.LottoNumber
+import lotto.service.Constants
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
