+++ src/test/kotlin/lotto/ManualTicketCountValidatorTest.kt
@@ -1,5 +1,6 @@
 package lotto
 
+import lotto.model.ManualTicket
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertDoesNotThrow
