+++ src/test/kotlin/lotto/TicketTest.kt
@@ -1,5 +1,8 @@
 package lotto
 
+import lotto.model.LottoNumber
+import lotto.model.Ticket
+import lotto.service.Constants
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertDoesNotThrow
