+++ src/test/kotlin/lotto/WinningNumbersTest.kt
@@ -1,5 +1,9 @@
 package lotto
 
+import lotto.model.LottoNumber
+import lotto.model.Ticket
+import lotto.model.WinningTicket
+import lotto.service.Constants
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
