+++ src/test/kotlin/lotto/WinningRankTest.kt
@@ -1,5 +1,6 @@
 package lotto
 
+import lotto.service.Rank
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
